requirements.txt
PyGithub
commonmark
beautifulsoup4
pydantic==1.9.0
